0.2.3 - Move sleep to optionalDependencies.
Mark sleep as an optional dependency to prevent module installation from
failing if for some reason installation of sleep fails (this uses a
compiled component).

In the worst case, a busy loop is used instead of a sleep if "sleep"
fails to install and the user uses new Xvfb(...).startSync().